+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- support for Toolbox 2.6.3 with improved URI handling
+
 ## 0.2.3 - 2025-05-26
 
 ### Changed