+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.3.0 - 2025-06-10
+
 ### Added
 
 - support for Toolbox 2.6.3 with improved URI handling